+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Version 2.0.0
 * AnalyzeLifeCycleNotifierAwareInterface to AnalyzeLifeCycleNotifierAware
 * AnalyzeLifeCycleNotifierInterface to LifeCycleNotifier
 * Analyzer to CoverageAnalyzer
+* AnalyzerInterface to Analyzer
 
 Version 1.8.3
 ---------------------------------------------------------------